# Improve performance of document creation.