--- a/lib/document.js
+++ b/lib/document.js
@@ -306,7 +306,7 @@ function $__applyDefaults(doc, fields, skipId, exclude, hasIncludedChildren, isB
     }
 
     const type = doc.schema.paths[p];
-    const path = p.split('.');
+    const path = p.indexOf('.') === -1 ? [p] : p.split('.');
     const len = path.length;
     let included = false;
     let doc_ = doc._doc;
@@ -982,7 +982,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
   }
 
   let schema;
-  const parts = path.split('.');
+  const parts = path.indexOf('.') === -1 ? [path] : path.split('.');
 
   // Might need to change path for top-level alias
   if (typeof this.schema.aliases[parts[0]] == 'string') {