--- a/lib/schema.js
+++ b/lib/schema.js
@@ -817,6 +817,9 @@ function _getPath(schema, path, cleanPath) {
  */
 
 function _pathToPositionalSyntax(path) {
+  if (!/\.\d+/.test(path)) {
+    return path;
+  }
   return path.replace(/\.\d+\./g, '.$.').replace(/\.\d+$/, '.$');
 }
 
@@ -1084,7 +1087,7 @@ Schema.prototype.indexedPaths = function indexedPaths() {
 
 Schema.prototype.pathType = function(path) {
   // Convert to '.$' to check subpaths re: gh-6405
-  const cleanPath = path.replace(/\.\d+\./g, '.$.').replace(/\.\d+$/, '.$');
+  const cleanPath = _pathToPositionalSyntax(path);
 
   if (this.paths.hasOwnProperty(path)) {
     return 'real';