**Summary**

Performance improvements for document creation. I used the already present `benchmarks/create.js` to validate these changes. 

The most common case for _pathToPositionalSyntax is that there's no positional syntax present. By avoiding the two calls to Regexp.replace, we can avoid creating two additional copies of the given path string per call. During document creation this function is called very frequently.

By avoiding extraneous calls to string split, we can reduce the number of string copies performed and garbage generated as a result.

**Examples**

On `benchmarks/create.js` locally:

Before:
```
creation of large object took 9617 ms
```
After:
```
creation of large object took 5320 ms, 5.32 ms per object
```
While the raw improvement doesn't seem large, on a larger document this takes the creation time from 170.11 ms per object to 90.56 ms, 80 ms per document created is a nice win.